+++ src/controllers/index.ts
@@ -1,6 +1,6 @@
 import { findTherapistById, getAllTherapists, updateTherapistProfile } from './therapists';
 import { getAppointments, updateAvailable, addAppointment } from './appointment';
-import login from './userLogin';
+import login from './auth';
 
 export {
   findTherapistById, getAllTherapists, getAppointments, updateAvailable,
