+++ src/services/index.ts
@@ -2,9 +2,9 @@ import { getTherapistById, getAllTherapist, updateTherapist } from './therapists
 import {
   getAppointmentsPerDateService, getAppointmentById, updateIsAvailable, addAppointment,
 } from './appointment';
-import LoginByEmail from './userLogin';
+import loginByEmail from './auth';
 
 export {
   getTherapistById, getAllTherapist, getAppointmentsPerDateService, getAppointmentById
-  , updateIsAvailable, addAppointment, updateTherapist, LoginByEmail,
+  , updateIsAvailable, addAppointment, updateTherapist, loginByEmail,
 };
